pyproject.toml
[project]
name = "dicom-server"
version = "0.1.0"
description = "FastAPI backend for the DICOM medical imaging platform (LATIS PFE)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "Oussama Ammari", email = "oussamaammari157@gmail.com" },
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Settings & validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.29.0",
    "alembic>=1.13.3",
    "psycopg2-binary>=2.9.9",  # for alembic sync migrations

    # Security
    "argon2-cffi>=23.1.0",
    "pyjwt[crypto]>=2.9.0",

    # Redis / caching
    "redis[hiredis]>=5.1.0",

    # Object storage
    "minio>=7.2.9",

    # Background tasks
    "celery[redis]>=5.4.0",

    # DICOM processing (our companion library)
    "pydicom>=2.4.4",
    # "medical-image-std @ git+https://github.com/LATIS-DocumentAI-Group/medical-image-std.git",

    # HTTP client
    "httpx>=0.27.2",

    # Observability
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.2",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "black>=24.10.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# ---------------------------------------------------------------------------
# Tooling
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.black]
line-length = 100
target-version = ["py311"]
exclude = "alembic/versions"

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["alembic/versions"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"
